3.5.7.9-leisure
Gridcoin Research 3.5.7.9/MSI=41.4
Mandatory Upgrade for TESTNET
Leisure Upgrade for PROD

- At block 196300 in TestNet, we require POR blocks to be signed (with
the CPID private key)
- After staking a testnet POR block, run the command "showblock
blocknumber" and verify the block signature is valid
- Test voting in TestNet
- Test Binary superblocks: Verify the superblock is staked, the block
number, and in showblock, verify the superblock is binary, and verify
the constiuent magnitude (and POR payment) is correct.
- Testnet Thread (please go here to continue testing):
https://cryptocointalk.com/topic/40141-testnet-research-age/page-34
- Fixed log spam in testnet

+++ src/rpcblockchain.cpp
@@ -482,8 +482,8 @@ Object blockToJSON(const CBlock& block, const CBlockIndex* blockindex, bool fPri
 		bool bIsBinary = Contains(bb.superblock,"<BINARY>");
 		result.push_back(Pair("IsBinary",bIsBinary));
 	}
-	result.push_back(Pair("IsSuperBlock", (double)blockindex->nIsSuperBlock));
-	result.push_back(Pair("IsContract", (double)blockindex->nIsContract));
+	result.push_back(Pair("IsSuperBlock", (int)blockindex->nIsSuperBlock));
+	result.push_back(Pair("IsContract", (int)blockindex->nIsContract));
     return result;
 }
